Move builtin definitions to a different file
